Flip the `index.tsx` with the `home-page.tsx` file
